% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_df_decimales.R
\name{my_df_decimales}
\alias{my_df_decimales}
\title{redondead un df dejando los decimales q le digas}
\usage{
my_df_decimales(df, nn = 2)
}
\arguments{
\item{df}{Un dataframe}

\item{nn}{el nº de decimales (Default = 2)}
}
\value{
Un nuevo \code{df} con sus valores redondeados a \code{nn}
}
\description{
decimales_df_pjp() te retorna un df con los decimales que le digas (default = 2)
decimales_df_pjp function redondea los valores de un dataframe. Le metes un df y redondea sus valores
}
\examples{
my_df_decimales(cars, nn = 2)
}
