% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_not_in.R
\name{my_not_in}
\alias{my_not_in}
\title{mi version de not_in}
\usage{
my_not_in(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
Un nuevo vector con los elemntos de x q no estan en y
}
\description{
te dice los valores de x q no estan en y
}
\examples{
aa <- my_not_in(x,y)
}
