% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pjp.funs-package.R
\docType{package}
\name{pjp.funs-package}
\alias{pjp.funs}
\alias{pjp.funs-package}
\title{pjp.funs: A collections of functions that are useful to me}
\description{
The package contains a group of functions that are useful to me.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/perezp44/pjp.funs}
  \item Report bugs at \url{https://github.com/perezp44/pjp.funs/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Perez \email{pedro.j.perez@uv.es}

}
\keyword{internal}
